Add on delete constraints

+++ data/src/main/java/me/polamokh/elcheck/data/local/order/Order.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "orders")
 data class Order(
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "order_id") val orderId: Int = 0,
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "order_id") val orderId: Long = 0,
     val name: String,
     @ColumnInfo(name = "order_time") val orderTime: String
 )
